refactor:revise message margin style
