+++ src/pages/Order/Order.tsx
@@ -1,6 +1,6 @@
 import Receiver from '@/components/Order/Receiver.tsx';
 import ItemInfo from '@/components/Order/ItemInfo.tsx';
-import Message from '@/components/Order/Message.tsx';
+import Message from '@/components/Order/Message/Message.tsx';
 import Sender from '@/components/Order/Sender.tsx';
 import Header from '@/components/Header/Header.tsx';
